svelte(phase 3): extract the git bar into GitFooter.svelte
Move the branch indicator + smart PR split button + branch rename out of the
imperative git-footer.js (called from ChatComposer.onMount) into a child
<GitFooter {sessionId} /> that owns the .pi-git-bar markup — including the
always-available btw button, still wired by <BtwPopup>. git-api.js stays as a
pure, injectable fetch helper.

Drops the dead dataset-sync first-paint path (the SPA never stamps data-git-*,
so it was inert; the bar is driven by the async getGitInfo). git-footer.js +
its test are removed (ported to GitFooter.test.js, 8 cases).
