+++ docs/dev/svelte-migration-plan.md
@@ -604,6 +604,12 @@ per increment.
 | ✅ DONE + verified | **`session.js` DELETED — head relocated into `<SessionPage>` + two utils.** The remaining `runSessionApp` head moved to `<SessionPage>`'s `onMount` as a local `startSessionRuntime()`: per-page settings/markdown bootstrap, `setupSessionUi`, the content-runtime wiring, the `__piIsMobileLayout`/`__piCloseSidebar` bridges, `attachHeaderHandlers`, the initial `navigateTo`, and the annotation-layer `init({...})`. Two extracted utils: **`lazy-highlight.js`** (`applyLazyHighlighting`) and **`session-content-runtime.js`** (`wireSessionContentRuntime`: builds the entry renderer + `sessionFormat`, assigns `contentRuntime.renderEntry`/`afterRender`, sets `downloadSessionJson`, and binds the single delegated copy/fork/label handler on `#messages`). **`session/session.js` + `session.test.js` DELETED** (the dead re-exports went with it; reconcile is unit-tested on the model). Go guard `TestNavigationReappliesCurrentToggleStateAfterRenderingMessages` repointed from `session.js` → `session-content-runtime.js`. Verified: web 535 + knip clean + build (no a11y) + `go test ./internal/ui` + go vet + **Desktop-Chrome e2e 57/2 (clean, no flake)**. |
 
 **`session.js` teardown COMPLETE** — the live session runtime is now `<SessionPage>` (orchestration) + `session-globals.js` + `session-content-runtime.js` + `lazy-highlight.js`, all reading the reactive `SessionDataModel`. The remaining §8 deletions (the chat/live runner modules still owned by `<ChatComposer>`/`<LiveReload>`, and decomposing `session-entry-renderer.js` into Svelte sub-components) are the next tranche.
+
+### Phase 3 — chat/live runner inlining (IN PROGRESS)
+
+| Status | Item |
+|---|---|
+| ✅ DONE + verified | **`GitFooter.svelte`** (chat tranche, first cut). Extracted the git bar (branch indicator + smart PR split button + branch rename) from `git-footer.js` (264 LOC, called directly in `ChatComposer.onMount`) into a child `<GitFooter {sessionId} />` that owns the `.pi-git-bar` markup — including the always-available `#pi-btw-button` (still wired by `<BtwPopup>`). `git-api.js` stays as a **pure** fetch helper (injectable `gitApi` prop, default = the real module). Dropped the dead **dataset-sync first-paint path** (the SPA never stamps `data-git-*`, so it was inert — the bar is driven by the async `getGitInfo`). **`chat/git-footer.js` + `git-footer.test.js` DELETED** (ported to `GitFooter.test.js`, 8 cases). Verified: web 534 + knip clean + no a11y (only the standard `{@html icon}` notes) + `go test ./internal/ui` + go vet + Desktop-Chrome e2e (chat/btw/composer/slash/mention green; full 57/2). |
 | 🔶 remaining | **Phase 2 was NOT isolatable from `session.js` (resolved for tree+header).** The tree's rendering, the active leaf/target (`currentLeafId`/`currentTargetId`), and `filterMode`/`searchQuery` all live as imperative locals in `session.js` (and `export-entry.js`); tree clicks drive **content** rendering via the navigator. Swapping only the tree DOM to `<SessionTreeNodes>` would need a throwaway bridge between that imperative state and the reactive model — which Phase 3 then deletes. **Do the tree cut-over together with moving navigation + filter state into `SessionDataModel`** (i.e. merge the front of Phase 3 into Phase 2), so the model is the single source of truth and `session.js`/`export-entry.js` stop owning that state. The staged `TreeNode`/`SessionTreeNodes` components are ready for that step. |
 
 **Recommended next step (combined Phase 2/3a):** move `currentLeafId`,
