+++ web/src/components/session/ChatComposer.svelte
@@ -1,6 +1,6 @@
 <script>
   import { onMount } from 'svelte';
-  import { icon, Maximize2, Paperclip, ChevronDown, ExternalLink, X } from '../../shared/icons.js';
+  import { icon, Maximize2, Paperclip, X } from '../../shared/icons.js';
   import { t } from '../../shared/i18n.js';
   import { escapeHtml } from '../../session/render/session-format.js';
   import { runChatComposer } from '../../session/chat/chat-composer-runner.js';
@@ -10,8 +10,7 @@
   import * as thinkingSelector from '../../session/chat/thinking-selector.js';
   import * as slashSelector from '../../session/chat/slash-command.js';
   import * as mentionSelector from '../../session/chat/mention-autocomplete.js';
-  import * as gitApi from '../../session/chat/git-api.js';
-  import { setupGitFooter } from '../../session/chat/git-footer.js';
+  import GitFooter from './GitFooter.svelte';
 
   let {
     sessionId = '',
@@ -50,8 +49,6 @@
       CustomEventImpl: target.CustomEvent,
       setIntervalImpl: target.setInterval.bind(target),
     });
-
-    setupGitFooter({ documentImpl: document, windowImpl: target, sessionId, gitApi });
   });
 </script>
 
@@ -120,5 +117,5 @@
       <div class="pi-chat-attachment-card-note" hidden></div>
     </div>
   </div>
-  <div class="pi-git-bar" id="pi-git-bar" data-git-repo="false" data-git-branch="" data-git-default="false" data-git-has-changes="false"><div class="pi-git-branch" id="pi-git-branch" hidden><span class="pi-git-branch-name" id="pi-git-branch-name" title={t('git.currentBranch')}></span><button type="button" class="pi-git-edit" id="pi-git-branch-edit" title={t('git.renameBranch')} aria-label={t('git.renameBranch')}></button><input type="text" class="pi-git-branch-input" id="pi-git-branch-input" autocomplete="off" spellcheck="false" aria-label={t('git.newBranchName')} hidden></div><div class="pi-git-right"><button type="button" class="pi-git-pr-button pi-btw-button" id="pi-btw-button" title="btw">btw</button><div class="pi-git-pr" id="pi-git-pr" hidden><button type="button" class="pi-git-pr-button pi-git-primary" id="pi-git-primary"><span id="pi-git-primary-label">{t('git.createPr')}</span></button><button type="button" class="pi-git-pr-button pi-git-caret" id="pi-git-caret" aria-haspopup="true" aria-expanded="false" aria-label={t('git.moreActions')}>{@html icon(ChevronDown, { size: 12 })}</button><div class="pi-git-pr-menu" id="pi-git-pr-menu" role="menu" hidden><button type="button" class="pi-git-pr-item" id="pi-git-pr-view" role="menuitem" hidden>{t('git.viewPr')} {@html icon(ExternalLink, { size: 12 })}</button><button type="button" class="pi-git-pr-item" id="pi-git-pr-draft" role="menuitem" hidden>{t('git.createDraftPr')}</button><button type="button" class="pi-git-pr-item" id="pi-git-pr-manual" role="menuitem">{t('git.createPrManually')} {@html icon(ExternalLink, { size: 12 })}</button><button type="button" class="pi-git-pr-item" id="pi-git-pr-merge" role="menuitem" hidden>{t('git.mergePr')}</button><button type="button" class="pi-git-pr-item" id="pi-git-pr-commit" role="menuitem" hidden>{t('git.commitPush')}</button></div></div></div></div>
+  <GitFooter {sessionId} />
 </form>
